feat: lazy loading of the init command
